% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hillshade.R
\name{choose_relief_colors}
\alias{choose_relief_colors}
\title{Choose relief colors for hillshade based on map background color}
\usage{
choose_relief_colors(bg_color)
}
\arguments{
\item{bg_color}{Background color in hex format}
}
\value{
List with light and dark relief colors
}
\description{
Choose relief colors for hillshade based on map background color
}
