% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_race_series.R
\name{create_race_series}
\alias{create_race_series}
\title{Loop over multiple races and styles to create maps}
\usage{
create_race_series(
  output_dir = "maps",
  styles_path = "styles.yaml",
  races_path = "races.yaml"
)
}
\arguments{
\item{output_dir}{Directory to save maps}

\item{styles_path}{Path to styles YAML}

\item{races_path}{Path to races YAML}
}
\value{
Saves maps for all races and styles
}
\description{
Loop over multiple races and styles to create maps
}
