% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hillshade.R
\name{geom_relief}
\alias{geom_relief}
\title{Hillshade and shaded relief functions for ggplot2}
\usage{
geom_relief(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  raster = TRUE,
  interpolate = TRUE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()} or \code{aes_()}.}

\item{data}{The data to be displayed in this layer. If NULL, the default
data frame will be used.}

\item{stat}{The statistical transformation to use on the data for this layer.
Default is "identity".}

\item{position}{Position adjustment for overlapping points. Default is "identity".}

\item{...}{Other arguments passed on to \code{layer()}. These are often aesthetics,
used to set an aesthetic to a fixed value, like \code{color = "red"} or \code{size = 3}.
They may also be parameters to the paired geom/stat.}

\item{raster}{Logical, whether to use raster rendering (default TRUE).}

\item{interpolate}{Logical, whether to interpolate raster (default TRUE).}

\item{na.rm}{Logical, whether to remove missing values (default FALSE).}

\item{show.legend}{Logical or NA, whether to include this layer in the legends.
Default is NA.}

\item{inherit.aes}{Logical, whether to inherit aesthetics from the plot mapping.
Default is TRUE.}
}
\value{
A ggplot2 layer for shaded relief plotting.
}
\description{
Provides geom_relief and supporting functions for shaded relief plots.
}
\seealso{
\code{\link{.derv}}, \code{\link{.rgb2hex}}, \code{\link{rect_to_poly}}
}
