% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_track_stats.R
\name{get_track_stats}
\alias{get_track_stats}
\title{Get max/min distance and elevation from multiple GPX files}
\usage{
get_track_stats(gpx_dir)
}
\arguments{
\item{gpx_dir}{Directory containing GPX files}
}
\value{
List with min and max distance and elevation
}
\description{
Get max/min distance and elevation from multiple GPX files
}
