% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gpx.R
\name{parse_gpx}
\alias{parse_gpx}
\title{Parse GPX file and extract track points}
\usage{
parse_gpx(gpx_file)
}
\arguments{
\item{gpx_file}{Path to the GPX file}
}
\value{
Data frame with latitude, longitude, elevation, and point order
}
\description{
Parse GPX file and extract track points
}
