% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_powerof10_to_yaml.R
\name{save_powerof10_to_yaml}
\alias{save_powerof10_to_yaml}
\title{Save Powerof10 athlete race data to YAML}
\usage{
save_powerof10_to_yaml(
  first_name,
  surname,
  club,
  yaml_path = NULL,
  event = NULL,
  year = NULL
)
}
\arguments{
\item{first_name}{Athlete's first name}

\item{surname}{Athlete's surname}

\item{club}{Athlete's club}

\item{yaml_path}{Path to save the YAML file (default: "<first_name>_\if{html}{\out{<surname>}}_races.yaml")}

\item{event}{Optional vector of events to filter e.g "HM", "Mar", "5K "parkrun" etc.}

\item{year}{Optional vector of years to filter}
}
\value{
Saves a YAML file with the athlete's race data
}
\description{
Save Powerof10 athlete race data to YAML
}
