% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osm_components.R
\name{get_osm_components}
\alias{get_osm_components}
\title{Get OSM map components (highways, streets, water, sea) for a bounding box and location}
\usage{
get_osm_components(
  bbox,
  location,
  cache_data = TRUE,
  components = c("highways", "streets", "water", "coast")
)
}
\arguments{
\item{bbox}{Numeric vector of length 4: c(min_lon, min_lat, max_lon, max_lat)}

\item{location}{String identifier for caching}

\item{cache_data}{Logical, whether to cache OSM data locally}

\item{components}{Character vector specifying which OSM components to include. Any combination of 'highways', 'streets', 'water', 'coast'. Defaults to all.}
}
\value{
List with selected OSM components
}
\description{
Get OSM map components (highways, streets, water, sea) for a bounding box and location
}
